ocr-service/requirements.txt
flask==3.0.0
paddleocr==2.7.0
paddlepaddle==2.5.2
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
